gio: Use new exiv2 type

+++ gexiv2/gexiv2-metadata.cpp
@@ -140,7 +140,7 @@ public:
         return -1;
     }
 
-    long size() const {
+    size_t size() const {
         return -1;
     }
